* changes:
  fix(bl31): add missing curly braces
  fix(xilinx): match function type as its declared
  fix(platforms): typedef operands to match data type
  fix(platforms): declare unused parameters as void
  fix(platforms): add essential bool type
  fix(platforms): fix misra violation 10.1

+++ plat/xilinx/common/plat_clkfunc.c
@@ -11,7 +11,7 @@
 #include <plat_clkfunc.h>
 #include <plat_private.h>
 
-uint32_t plat_get_syscnt_freq2(void)
+unsigned int plat_get_syscnt_freq2(void)
 {
 	uint32_t counter_freq = 0;
 	uint32_t ret = 0;
@@ -26,7 +26,7 @@ uint32_t plat_get_syscnt_freq2(void)
 		ret = cpu_clock;
 	}
 
-	return ret;
+	return (unsigned int)ret;
 }
 
 void set_cnt_freq(void)
